requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
defusedxml==0.7.1
Django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
Pygments==2.18.0
PySocks==1.7.1
python-slugify==8.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
urllib3==2.2.2
whitenoise==6.7.0
